refactor: move settings-matrix parent-run persistence to core
