+++ homeboy.json
@@ -680,7 +680,6 @@
         "test_quality::src/core/runner/execution/tests/redaction.rs::VacuousTest",
         "test_quality::src/core/runner/execution/tests/secret_source.rs::VacuousTest",
         "test_quality::tests/command_surface_test.rs::VacuousTest",
-        "thin_command_adapter::src/commands/bench/settings_matrix.rs::ThinCommandAdapterViolation",
         "thin_command_adapter::src/commands/observe.rs::ThinCommandAdapterViolation",
         "thin_command_adapter::src/commands/review/mod.rs::ThinCommandAdapterViolation",
         "thin_command_adapter::src/commands/runs/gh_actions.rs::ThinCommandAdapterViolation",
